refactor(cef): split browser_panes.rs into browser_panes/ submodules
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
BrowserPaneManager is a zero-field unit struct — every method takes
state: &Arc<AppState> as a parameter, so its ~913-line impl block splits
across files with zero shared-state risk. Split browser_panes/{mod,close,
navigation,zoom,clip}.rs by method cluster (creation, close/lifecycle,
navigation, zoom, overlay-clip+focus). compute_pane_visible is re-exported
from mod.rs (cfg-gated to match its definition) since browser_pane/
creation_views.rs calls it externally via crate::browser_panes::
compute_pane_visible.

Pure relocation, no logic changes. cargo build -p agentmux-cef clean;
cargo test -p agentmux-cef browser_panes:: — 7/7 pass, same as before.

<!-- agentmux:agent_id=agentx -->
