--- agentmux-cef/src/browser_panes.rs
@@ -1,1480 +1,0 @@
-// Copyright 2026, AgentMux Corp.
-// SPDX-License-Identifier: Apache-2.0
-
-//! BrowserPaneManager: embeds browsers as native OS child windows using
-//! CefBrowserHost::CreateBrowser. All creation runs on the CEF UI thread.
-//!
-//! The Browser instance is owned by the host reducer's `browsers` map (keyed
-//! by label, accessed via `AppState::get_browser` etc). We only need the
-//! block_id → label mapping, which also lives in the reducer's `panes` map.
-//!
-//! Lifecycle states are tracked explicitly via the reducer's `BrowserPaneLifecycle`:
-//!   Created → Closing → Closed (removed from `panes`)
-//! Every pane-facing op (focus/resize/navigate/…) short-circuits when the
-//! entry is already in `Closing`. This drops late IPC that the frontend
-//! fires after it has already asked for close but before CEF has destroyed
-//! the Browser — stale IPC against a mid-destruction HWND is the shape of
-//! the crash described in `docs/specs/SPEC_BROWSER_PANE_LIFECYCLE.md` §4c.
-//!
-//! **Phase H.1.d/e (PR #5):** The legacy `pane::lifecycle::PaneStateMachine`
-//! is gone; pane state lives only in `HostState.browser_panes`. All mutations go
-//! through `HostCommand::TryRegisterBrowserPaneLive` / `EnqueueBrowserPaneClose` /
-//! `CompleteBrowserPaneClose` / `DrainBrowserPaneByLabel` and read back via the reducer's
-//! atomic `DispatchOutput` fields.
-
-use std::sync::Arc;
-
-use cef::*;
-
-use crate::browser_pane::CreateBrowserPaneTask;
-use crate::reducer::RegisterResult;
-use crate::state::AppState;
-
-/// Abstraction over the CEF-side operations `close()` performs on the host
-/// process. Production implements this over `&Arc<AppState>` and Win32;
-/// tests implement it with a recording mock so the close-path state machine
-/// can be exercised without real CEF/HWNDs.
-///
-/// Kept minimal (just two methods) to avoid a dependency graph that has to
-/// be updated every time `close()` grows. Other ops (`focus`, `resize`, …)
-/// can gain their own traits when they need testing, or graduate to a
-/// unified `BrowserPaneCefBridge` once the shape is stable.
-pub trait BrowserPaneCloseOps {
-    /// Remove the Browser for this label from the registry. Return its
-    /// outer HWND as a pointer-sized value, or `None` if there is no
-    /// Browser or no HWND. Dropping the Browser Arc is the implementation's
-    /// responsibility — production drops before returning so Chromium's
-    /// refcount isn't held by our scope.
-    fn take_browser_hwnd(&self, label: &str) -> Option<usize>;
-
-    /// Destroy the given HWND. Production posts a CEF UI-thread task
-    /// (`DestroyPaneWrapperTask`) that runs Win32 `DestroyWindow` on the
-    /// wrapper's OWNING thread — DestroyWindow is owner-thread-only, and
-    /// calling it directly from the IPC thread silently no-ops
-    /// (retro-browser-pane-renderer-leak-2026-07-07). Called only with
-    /// values returned from `take_browser_hwnd`.
-    fn destroy_hwnd(&self, hwnd: usize);
-}
-
-/// Production implementation of `BrowserPaneCloseOps` backed by `AppState.browsers`
-/// and Win32 `DestroyWindow`.
-///
-/// SPEC_BROWSER_PANE_WINDOWS_TEARDOWN_SPIKE_2026_07_03.md: `take_browser_hwnd`
-/// now returns the app-owned WRAPPER's HWND (`browser_pane::wrapper`), not
-/// CEF's own HWND — `destroy_hwnd` destroys the wrapper (after reparenting it
-/// out to top-level; see `destroy_wrapper_hwnd`'s doc for why that step is
-/// what actually releases the renderer — retro-browser-pane-renderer-leak-
-/// 2026-07-07), never CEF's own window directly and never via
-/// `close_browser()`, avoiding the close_browser-cascades-into-main bug three
-/// earlier attempts hit.
-struct AppStateCloseOps<'a>(&'a Arc<AppState>);
-
-impl<'a> BrowserPaneCloseOps for AppStateCloseOps<'a> {
-    fn take_browser_hwnd(&self, label: &str) -> Option<usize> {
-        // Atomic take-and-return via reducer (codex P2 PR #660). Earlier
-        // round 1 separated `get_browser` + `UnregisterBrowser` dispatch,
-        // which left a window for concurrent readers to resolve the same
-        // label and act on the closing handle. `UnregisterBrowser` now
-        // returns the removed `Browser` via `DispatchOutput.removed_browser`
-        // — single host_state lock, single mutation, no race.
-        let out = self.0.host_dispatch(
-            crate::reducer::HostCommand::UnregisterBrowser {
-                label: label.to_string(),
-            },
-        );
-        let browser = out.removed_browser?;
-
-        #[cfg(target_os = "windows")]
-        {
-            // Forget any stale focus-tracking entry for CEF's OWN hwnd
-            // (LAST_FOCUSED_BY_ROOT is keyed by whatever HWND actually
-            // received SetFocus — always CEF's own hwnd/descendants, never
-            // our wrapper, which never itself receives focus). Belt-and-
-            // suspenders: `on_before_close_browser_pane` also runs a more
-            // thorough sweep via `uninstall_focus_redirect_for_block` once
-            // CEF's OnBeforeClose fires (which the wrapper's reparent-then-
-            // destroy teardown is what makes reliable — see
-            // destroy_wrapper_hwnd) — this covers the gap before that.
-            if let Some(host) = browser.host() {
-                let wh = host.window_handle();
-                if !wh.0.is_null() {
-                    crate::browser_pane::hwnd::forget_focus_for_child(wh.0 as *mut _);
-                }
-            }
-        }
-
-        // Drop our Arc before returning so Chromium's refcount doesn't wait
-        // for the caller's scope to unwind.
-        drop(browser);
-
-        #[cfg(target_os = "windows")]
-        {
-            crate::browser_pane::wrapper::take_wrapper_hwnd(label).map(|h| h as usize)
-        }
-        #[cfg(not(target_os = "windows"))]
-        {
-            None
-        }
-    }
-
-    fn destroy_hwnd(&self, hwnd: usize) {
-        #[cfg(target_os = "windows")]
-        {
-            // Marshal the destroy onto the CEF UI thread — the thread that
-            // CREATED the wrapper (CreateBrowserPaneTask::execute). `close()`
-            // runs on a tokio IPC thread, and Win32's DestroyWindow hard-fails
-            // (ERROR_ACCESS_DENIED) from any thread other than the window's
-            // owner — which is exactly how every pane close silently leaked
-            // its renderer while the logs looked clean
-            // (retro-browser-pane-renderer-leak-2026-07-07: the pixels
-            // vanished because ShowWindow(SW_HIDE) works cross-thread; the
-            // DestroyWindow after it never did anything, so CEF never saw
-            // WM_DESTROY and the browser survived headless). Mirrors the
-            // Linux/macOS path, which has always posted
-            // DetachBrowserPaneViewTask for the same stated reason (see the
-            // marshalling-tasks comment near the bottom of this file).
-            let mut task = DestroyPaneWrapperTask::new(hwnd as isize);
-            let posted = cef::post_task(cef::ThreadId::UI, Some(&mut task));
-            if posted == 0 {
-                tracing::error!(
-                    hwnd,
-                    "[pane-wrapper] post_task(destroy) failed — wrapper + CEF browser will leak"
-                );
-            }
-        }
-        #[cfg(not(target_os = "windows"))]
-        {
-            let _ = hwnd;
-        }
-    }
-}
-
-/// Signature of the clip a pane should have, used by `set_pane_overlay_clip`
-/// to skip a redundant `SetWindowRgn` when nothing changed. The HWND value is
-/// folded in so a recreated pane (new HWND under a reused label) never matches a
-/// stale entry — see `AppState::pane_clip_cache`. Within-process only (the hash
-/// need not be stable across runs).
-#[cfg(target_os = "windows")]
-fn clip_sig_whole(hwnd: isize) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut h = std::collections::hash_map::DefaultHasher::new();
-    0u8.hash(&mut h); // tag: full visibility (region = NULL)
-    hwnd.hash(&mut h);
-    h.finish()
-}
-
-#[cfg(target_os = "windows")]
-fn clip_sig_clipped(
-    hwnd: isize,
-    pane_left: i32,
-    pane_top: i32,
-    pane_w: i32,
-    pane_h: i32,
-    overlay_rects: &[(i32, i32, i32, i32)],
-) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut h = std::collections::hash_map::DefaultHasher::new();
-    1u8.hash(&mut h); // tag: clipped
-    hwnd.hash(&mut h);
-    // The applied region is a deterministic function of the pane's screen
-    // position/size and the overlay rects, so hashing those inputs is a faithful
-    // signature: identical inputs ⇒ identical region (no false cache hit). A
-    // non-intersecting rect that changes only yields a false MISS (a harmless
-    // redundant re-apply), never a false hit.
-    pane_left.hash(&mut h);
-    pane_top.hash(&mut h);
-    pane_w.hash(&mut h);
-    pane_h.hash(&mut h);
-    overlay_rects.hash(&mut h);
-    h.finish()
-}
-
-pub struct BrowserPaneManager;
-
-impl BrowserPaneManager {
-    pub fn new() -> Self {
-        Self
-    }
-
-    /// Look up the Browser iff the pane is Live. Returns None when closing
-    /// so all ops short-circuit uniformly.
-    fn live_browser(&self, state: &Arc<AppState>, block_id: &str) -> Option<Browser> {
-        let label = state.live_browser_pane_label(block_id)?;
-        state.get_browser(&label)
-    }
-
-    /// Return the current URL of the pane's main frame, if the pane
-    /// is Live. Used by the browser DOM API resolver
-    /// (`crate::browser_api::resolver`) to match CEF `/json` targets
-    /// against block ids without a first-class `browserId` field on
-    /// the CEF side.
-    pub fn pane_url(&self, state: &Arc<AppState>, block_id: &str) -> Option<String> {
-        let browser = self.live_browser(state, block_id)?;
-        let frame = browser.main_frame()?;
-        Some(CefString::from(&frame.url()).to_string())
-    }
-
-    pub fn create(
-        &self,
-        state: &Arc<AppState>,
-        block_id: &str,
-        url: &str,
-        rect: Rect,
-        window_label: &str,
-    ) -> Result<(), String> {
-        // Phase H.1.d (PR #5) — sole pane-registration entry point. The
-        // reducer atomically generates the label and inserts the entry,
-        // returning Fresh / AlreadyLive / Closing via DispatchOutput.
-        //
-        // We pass the create params as `pending`: if the reducer finds the
-        // block_id still `Closing`, it stashes them (under the same host_state
-        // lock that observed Closing — atomic, no TOCTOU) for the
-        // close-completion arm to replay. Ignored for Fresh/AlreadyLive.
-        let out = state.host_dispatch(
-            crate::reducer::HostCommand::TryRegisterBrowserPaneLive {
-                block_id: block_id.to_string(),
-                pending: Some(crate::state::PendingBrowserPaneCreate {
-                    url: url.to_string(),
-                    x: rect.x,
-                    y: rect.y,
-                    width: rect.width,
-                    height: rect.height,
-                    window_label: window_label.to_string(),
-                }),
-            },
-        );
-        let result = out.browser_pane_register_result.ok_or_else(|| {
-            format!(
-                "try_register_browser_pane_live returned no result (block_id={}); host shutting down?",
-                block_id
-            )
-        })?;
-        match result {
-            RegisterResult::AlreadyLive(label) => {
-                // Existing Live entry — re-navigate the existing browser.
-                //
-                // DIAGNOSTIC (redock black-page triage —
-                // ANALYSIS_BROWSER_PANE_REDOCK_BLACK_TYPING_LOCK_2026_06_15.md §2):
-                // if a redock's create lands here while the old pane is still
-                // Live (the close IPC hasn't flipped it to Closing yet), we
-                // only re-navigate the EXISTING browser — which is still
-                // parented to the *old* (floater) window — and never create a
-                // pane in the requested `window_label`. The target window then
-                // shows black until the old pane closes. Logging requested
-                // window + rect makes this race visible in a repro; if a redock
-                // hits this arm, it is the smoking gun.
-                tracing::info!(
-                    block_id,
-                    label = %label,
-                    requested_window = %window_label,
-                    rect = ?(rect.x, rect.y, rect.width, rect.height),
-                    "browser pane create hit AlreadyLive — re-navigating existing browser (NOT creating in requested window)"
-                );
-                crate::browser_pane::trace::pane_trace(
-                    block_id,
-                    "create-already-live",
-                    &format!("label={label} requested_win={window_label} url={url}"),
-                );
-                if let Some(browser) = state.get_browser(&label) {
-                    if let Some(frame) = browser.main_frame() {
-                        frame.load_url(Some(&CefString::from(url)));
-                    }
-                }
-                Ok(())
-            }
-            RegisterResult::AlreadyLiveElsewhere(label) => {
-                // Cross-window move (tear-off / redock): the block is Live in a
-                // DIFFERENT window. The reducer stashed our pending create;
-                // close the old pane now. Its close-completion (CompleteBrowserPaneClose
-                // / DrainBrowserPaneByLabel) drains the entry and replays the
-                // stashed create as `Fresh` in the requested window — so the
-                // browser reappears in the new (floating) window instead of the
-                // requested window rendering black. See
-                // ANALYSIS_BROWSER_PANE_REDOCK_BLACK_TYPING_LOCK_2026_06_15 §2.
-                tracing::info!(
-                    block_id,
-                    old_label = %label,
-                    requested_window = %window_label,
-                    rect = ?(rect.x, rect.y, rect.width, rect.height),
-                    "browser pane create is a cross-window move — closing old pane; reducer will replay create in requested window"
-                );
-                crate::browser_pane::trace::pane_trace(
-                    block_id,
-                    "create-cross-window-move",
-                    &format!("old_label={label} requested_win={window_label}"),
-                );
-                self.close(block_id, state);
-                Ok(())
-            }
-            RegisterResult::Closing => {
-                // Old CEF Browser mid-teardown — don't overwrite (its
-                // on_before_close → DrainBrowserPaneByLabel would evict the NEW
-                // entry), and don't drop the request. The reducer already
-                // stashed our `pending` create (above) under the host_state
-                // lock; the close-completion arm replays it once the old entry
-                // is gone (see `drain_closed_label` / `close()`). This is the
-                // redock case: the target window re-creates the same block_id
-                // while the floater's pane is still Closing. (The old "frontend
-                // retries on next tick" never existed — browser-view.tsx::createPane
-                // errored out — which is why redocked panes intermittently
-                // never loaded.) See
-                // docs/analysis/ANALYSIS_BROWSER_PANE_REDOCK_LOAD_RACE_2026_05_29.md.
-                tracing::info!(
-                    block_id,
-                    requested_window = %window_label,
-                    rect = ?(rect.x, rect.y, rect.width, rect.height),
-                    "browser pane create deferred — block_id still Closing; reducer will replay on close-completion"
-                );
-                crate::browser_pane::trace::pane_trace(
-                    block_id,
-                    "create-deferred-closing",
-                    "old entry still Closing; reducer will replay on close-completion",
-                );
-                Ok(())
-            }
-            RegisterResult::Fresh(label) => {
-                crate::browser_pane::trace::pane_trace(
-                    block_id,
-                    "create-request",
-                    &format!(
-                        "url={url} label={label} win={window_label} rect=({},{},{},{})",
-                        rect.x, rect.y, rect.width, rect.height
-                    ),
-                );
-                let mut task = CreateBrowserPaneTask::new(
-                    state.clone(),
-                    block_id.to_string(),
-                    label,
-                    url.to_string(),
-                    rect,
-                    window_label.to_string(),
-                );
-                post_task(ThreadId::UI, Some(&mut task));
-                Ok(())
-            }
-        }
-    }
-
-    pub fn navigate(&self, block_id: &str, url: &str, state: &Arc<AppState>) -> Result<(), String> {
-        if let Some(browser) = self.live_browser(state, block_id) {
-            if let Some(frame) = browser.main_frame() {
-                frame.load_url(Some(&CefString::from(url)));
-            }
-        }
-        Ok(())
-    }
-
-    pub fn resize(&self, block_id: &str, rect: Rect, state: &Arc<AppState>) {
-        // Windows: resize the wrapper (its own WM_SIZE handler cascades to
-        // CEF's child automatically — see browser_pane::wrapper) instead of
-        // SetWindowPos-ing CEF's HWND directly. Still need the live Browser
-        // to call notify_move_or_resize_started (a CEF-side hint, unrelated
-        // to which HWND owns the actual on-screen rect).
-        #[cfg(target_os = "windows")]
-        if let Some(browser) = self.live_browser(state, block_id) {
-            let resized = state
-                .live_browser_pane_label(block_id)
-                .and_then(|label| crate::browser_pane::wrapper::peek_wrapper_hwnd(&label))
-                .map(|wrapper_hwnd| {
-                    crate::browser_pane::wrapper::resize_wrapper(
-                        wrapper_hwnd as *mut std::ffi::c_void,
-                        &rect,
-                    );
-                })
-                .is_some();
-            // Only notify CEF if we actually resized something — matches the
-            // original hwnd-null gating this replaced.
-            if resized {
-                if let Some(host) = browser.host() {
-                    host.notify_move_or_resize_started();
-                }
-            }
-        }
-        // Linux/macOS — Views path. The pane is a CefBrowserView in the main
-        // window's view hierarchy; resizing is `View::set_bounds` in DIP. Must
-        // run on the CEF UI thread (set_bounds is UI-thread-only).
-        #[cfg(not(target_os = "windows"))]
-        {
-            let label = match state.live_browser_pane_label(block_id) {
-                Some(l) => l,
-                None => return, // pane already closed or never created
-            };
-            let mut task = ResizeBrowserPaneViewTask::new(state.clone(), label, rect);
-            cef::post_task(cef::ThreadId::UI, Some(&mut task));
-        }
-    }
-
-    /// Close a pane by destroying its app-owned wrapper HWND (not CEF's own
-    /// HWND, and never via `close_browser()`) and dropping the Browser Arc.
-    ///
-    /// We deliberately do **not** call `host.close_browser(force)`. Empirically
-    /// (host-log trace v0.33.251 and v0.33.252 in `SPEC_BROWSER_PANE_LIFECYCLE.md`
-    /// §4), CEF Alloy treats the pane Browser and the main Browser as a single
-    /// close unit when the pane's outer HWND is a child of main's top-level:
-    /// `close_browser(pane)` fires `do_close` on main too. Previous attempts
-    /// (force=0, force=1, a cascade-guard cancelling main's do_close) either
-    /// quit the whole app or orphaned the pane's pixels while blocking the
-    /// pane's own teardown.
-    ///
-    /// As of `SPEC_BROWSER_PANE_WINDOWS_TEARDOWN_SPIKE_2026_07_03.md` (and
-    /// corrected per retro-browser-pane-renderer-leak-2026-07-07), instead:
-    /// 1. Remove the Browser from `state.browsers` so subsequent lookups miss.
-    /// 2. Reparent our app-owned wrapper HWND (`browser_pane::wrapper`) out to
-    ///    a genuine top-level window, THEN Win32 `DestroyWindow` it — never
-    ///    CEF's own HWND directly, and never `close_browser()`. The reparent
-    ///    is load-bearing: destroying the wrapper while it was still a
-    ///    `WS_CHILD` of main delivered `WM_DESTROY` via parent-hierarchy
-    ///    tear-down, which CEF does NOT treat as a browser close — the
-    ///    browser/renderer silently survived every close (live-confirmed via
-    ///    CDP). Made top-level first, the destroy is structurally identical
-    ///    to the floater's proven teardown and CEF runs its real close
-    ///    pipeline. See `destroy_wrapper_hwnd`'s doc for the full mechanism.
-    /// 3. Drop our `Browser` Arc. `on_before_close` fires via the
-    ///    reparent-then-destroy teardown; `drain_closed_label` stays
-    ///    idempotent as a backstop for any remaining async-timing edge case.
-    ///
-    /// Trade-off: because we bypass `close_browser`, Chromium's `beforeunload`
-    /// handler doesn't run. Acceptable for a browser pane (no form data the
-    /// user expects to persist across close). If beforeunload becomes
-    /// important, revisit.
-    pub fn close(&self, block_id: &str, state: &Arc<AppState>) {
-        crate::browser_pane::trace::pane_trace(block_id, "close", "");
-        // Phase H.1.d (PR #5) — sole pane-close entry point. The reducer
-        // flips Live→Closing atomically and returns the entry's label iff
-        // the transition fired. None means missing or already-Closing —
-        // both idempotent no-ops; we don't dispatch CompleteBrowserPaneClose in
-        // those cases (codex P2 PR #655 race), avoiding the entry removal
-        // while another in-flight close is still tearing down the HWND.
-        let close_out = state.host_dispatch(
-            crate::reducer::HostCommand::EnqueueBrowserPaneClose {
-                block_id: block_id.to_string(),
-            },
-        );
-        let label = match close_out.closed_browser_pane_label {
-            Some(l) => l,
-            None => return,
-        };
-        #[cfg(target_os = "windows")]
-        {
-            let ops = AppStateCloseOps(state);
-            Self::close_with(&label, &ops);
-            // Keyboard-focus orphaning fix: destroying the pane's native HWND
-            // doesn't hand Win32 focus back, so if the pane held focus (common
-            // on redock — the dragged pane is focused) keystrokes are orphaned
-            // app-wide until something reclaims focus. Post a focus reclaim for
-            // the surviving foreground window so its render widget (which hosts
-            // the whole DOM UI — agent/terminal panes included) regains focus
-            // and native panes are defocused. Mirrors the manual recovery the
-            // user otherwise triggers by switching windows. See
-            // ANALYSIS_BROWSER_PANE_REDOCK_BLACK_TYPING_LOCK_2026_06_15.md §1.
-            Self::reclaim_focus_after_pane_destroy(state);
-        }
-        // Linux/macOS — Views path. Marshal the BrowserView detach onto the
-        // CEF UI thread (remove_child_view is UI-thread-only); the underlying
-        // Browser's on_before_close fires asynchronously and clears
-        // state.browsers via the existing callback (callbacks::on_before_close_browser_pane).
-        #[cfg(not(target_os = "windows"))]
-        {
-            let mut task = DetachBrowserPaneViewTask::new(state.clone(), label.clone());
-            cef::post_task(cef::ThreadId::UI, Some(&mut task));
-        }
-        let close_out = state.host_dispatch(
-            crate::reducer::HostCommand::CompleteBrowserPaneClose {
-                block_id: block_id.to_string(),
-            },
-        );
-        tracing::info!(block_id, label, "browser pane closed");
-
-        // The explicit-close path removes the reducer entry here (not via the
-        // async drain), so it must ALSO replay any create deferred while this
-        // block_id was Closing — otherwise the deferred create is orphaned
-        // (pane never loads + leak). The reducer removed the stash atomically
-        // and handed it back here. reagent P1/P2 on PR #1168.
-        self.replay_pending_create(state, close_out.pending_browser_pane_create_to_replay);
-
-        // PR #6 H.7 kick — top up the pool now that this pane has closed.
-        // `spawn_pool_window` is internally idempotent (single-flight +
-        // below-target check), so calling on every pane close is safe.
-        //
-        // Cross-platform: the original `weak_ptr.h:250` race that prompted
-        // an earlier Windows-only cfg-gate is gone. With the deferred
-        // OverlayController destroy (see
-        // browser_pane/creation_views.rs::detach_browser_pane_view),
-        // close() no longer destroys the controller synchronously — it
-        // just stashes it for on_before_close to destroy later. Creating
-        // a new pool window here therefore can't race a synchronous
-        // destroy of the just-closed pane's View. drain_closed_label's
-        // pool kick can't be relied on as the sole refill source either:
-        // CompleteBrowserPaneClose (dispatched above) already removed the
-        // reducer entry, so DrainBrowserPaneByLabel inside
-        // drain_closed_label is a no-op and never reaches its
-        // spawn_pool_window() call (codex P2 on PR #788).
-        crate::commands::window_pool::spawn_pool_window(state);
-    }
-
-    /// The testable side-effect body of `close()`. Given a pane's `label`,
-    /// remove its Browser handle and destroy its HWND. The state-machine
-    /// transition (Live→Closing) and the entry removal (CompleteBrowserPaneClose)
-    /// happen in `close()` via reducer dispatch — `close_with` is purely
-    /// the FFI side-effects that follow.
-    fn close_with(label: &str, ops: &dyn BrowserPaneCloseOps) {
-        if let Some(hwnd) = ops.take_browser_hwnd(label) {
-            ops.destroy_hwnd(hwnd);
-            tracing::info!(label, "pane wrapper destroy dispatched");
-        }
-    }
-
-    /// Post a focus reclaim for the foreground window after a pane HWND was
-    /// destroyed, so Win32 keyboard focus returns to that window's main render
-    /// widget (which hosts the entire DOM UI — agent/terminal panes included)
-    /// and native panes are defocused. Windows-only: the orphaning is a Win32
-    /// native-child-window issue. The empty label tells `MainFocusReclaimTask`
-    /// to resolve the foreground agentmux window itself, which is correct for
-    /// both redock (the floater is gone → the target window is foreground) and
-    /// an in-window close (that window stays foreground).
-    #[cfg(target_os = "windows")]
-    fn reclaim_focus_after_pane_destroy(state: &Arc<AppState>) {
-        let mut task = crate::ui_tasks::MainFocusReclaimTask::new(state.clone(), String::new());
-        cef::post_task(cef::ThreadId::UI, Some(&mut task));
-    }
-
-    /// Called from CEF's `on_before_close` if/when it fires for a pane
-    /// browser. The explicit `close()` path usually clears the entry first,
-    /// so this is a no-op in that case — but `on_before_close` may still
-    /// fire async as Chromium's refcount hits zero, and `DrainBrowserPaneByLabel`
-    /// is idempotent so the callback is safe.
-    ///
-    /// Also destroys the app-owned wrapper HWND (`browser_pane::wrapper`) if
-    /// one is still registered for this label — reagent P1 on PR #1957: if
-    /// CEF's `OnBeforeClose` fires WITHOUT `close()` having run first (a
-    /// crash, or something else calling `close_browser()` directly on this
-    /// pane), CEF tears down its own child HWND on its own initiative, but
-    /// that never touches OUR wrapper (destroying a child never destroys its
-    /// parent) — the wrapper would otherwise survive as a permanently
-    /// orphaned, childless window with nothing left to ever clean it up,
-    /// since the only other destroy site is `close_with` in the explicit
-    /// path. `take_wrapper_hwnd` is a no-op `None` when `close()` already
-    /// destroyed it, keeping this idempotent like the rest of the function.
-    pub fn drain_closed_label(&self, state: &Arc<AppState>, label: &str) {
-        #[cfg(target_os = "windows")]
-        if let Some(wrapper_hwnd) = crate::browser_pane::wrapper::take_wrapper_hwnd(label) {
-            crate::browser_pane::wrapper::destroy_wrapper_hwnd(wrapper_hwnd as *mut std::ffi::c_void);
-        }
-
-        let out = state.host_dispatch(
-            crate::reducer::HostCommand::DrainBrowserPaneByLabel {
-                label: label.to_string(),
-            },
-        );
-        if let Some(block_id) = out.drained_browser_pane_block_id {
-            tracing::info!(label, block_id = %block_id, "browser pane drained via on_before_close");
-            // Same keyboard-focus orphaning fix as close() — the async
-            // on_before_close drain also destroys the pane HWND.
-            #[cfg(target_os = "windows")]
-            Self::reclaim_focus_after_pane_destroy(state);
-            // PR #6 H.7 kick — see `close()` for rationale. The
-            // on_before_close path is the async drain; pool refill that
-            // was deferred while the pane was Closing should now resume.
-            crate::commands::window_pool::spawn_pool_window(state);
-
-            // Deterministic redock re-create: the reducer removed the stash
-            // atomically and handed back any create deferred while this
-            // block_id was Closing. Replay it (now Fresh).
-            self.replay_pending_create(state, out.pending_browser_pane_create_to_replay);
-        }
-    }
-
-    /// Replay a browser-pane create that the reducer deferred while `block_id`
-    /// was `Closing` and handed back on close-completion (via
-    /// `DispatchOutput.pending_browser_pane_create_to_replay`). Called from
-    /// BOTH close-completion paths — the async `drain_closed_label`
-    /// (`DrainBrowserPaneByLabel`) and the explicit `close()`
-    /// (`CompleteBrowserPaneClose`). No-op if nothing was deferred. The old
-    /// entry is gone by now, so the replayed `create` gets `Fresh` and posts
-    /// the `CreateBrowserPaneTask`. The stash lived in (and was removed from)
-    /// the reducer's `HostState` under the host_state lock — no separate map,
-    /// no TOCTOU.
-    fn replay_pending_create(
-        &self,
-        state: &Arc<AppState>,
-        pending: Option<(String, crate::state::PendingBrowserPaneCreate)>,
-    ) {
-        if let Some((block_id, p)) = pending {
-            tracing::info!(block_id, "replaying deferred browser pane create after close");
-            let rect = Rect { x: p.x, y: p.y, width: p.width, height: p.height };
-            if let Err(e) = self.create(state, &block_id, &p.url, rect, &p.window_label) {
-                tracing::warn!(block_id, error = %e, "deferred browser pane create replay failed");
-            }
-        }
-    }
-
-    pub fn go_back(&self, block_id: &str, state: &Arc<AppState>) {
-        if let Some(mut b) = self.live_browser(state, block_id) { b.go_back(); }
-    }
-    pub fn go_forward(&self, block_id: &str, state: &Arc<AppState>) {
-        if let Some(mut b) = self.live_browser(state, block_id) { b.go_forward(); }
-    }
-    pub fn reload(&self, block_id: &str, state: &Arc<AppState>) {
-        if let Some(mut b) = self.live_browser(state, block_id) { b.reload(); }
-    }
-
-    /// Ctrl+Wheel step, in the same +/-0.05 units and 0.5-2.0 clamp range
-    /// `frontend/app/store/zoom.win32.ts`'s WHEEL_STEP uses for every other
-    /// pane type, for a consistent feel even though this path never touches
-    /// that frontend module (browser-pane content is unreachable from the
-    /// DOM — see the module doc on `AppState::browser_pane_zoom`).
-    const ZOOM_STEP: f64 = 0.05;
-    const ZOOM_MIN: f64 = 0.5;
-    const ZOOM_MAX: f64 = 2.0;
-
-    pub fn zoom_in(&self, block_id: &str, state: &Arc<AppState>) {
-        self.step_zoom(block_id, state, Self::ZOOM_STEP);
-    }
-    pub fn zoom_out(&self, block_id: &str, state: &Arc<AppState>) {
-        self.step_zoom(block_id, state, -Self::ZOOM_STEP);
-    }
-
-    fn step_zoom(&self, block_id: &str, state: &Arc<AppState>, delta: f64) {
-        let new_factor = {
-            let mut map = state.browser_pane_zoom.lock();
-            let current = *map.get(block_id).unwrap_or(&1.0);
-            let next = next_zoom_factor(current, delta);
-            map.insert(block_id.to_string(), next);
-            next
-        };
-        self.apply_zoom(block_id, state, new_factor);
-    }
-
-    /// Inject the pane's current zoom factor as CSS `zoom` via
-    /// `execute_java_script` on its main frame. Deliberately not Chromium's
-    /// native `SetZoomLevel` — that's scoped by `HostZoomMap`, shared across
-    /// every browser pane on the same host/RequestContext (the bug this
-    /// exists to fix). CSS injection is per-`CefFrame` by construction, no
-    /// native zoom system involved at all.
-    fn apply_zoom(&self, block_id: &str, state: &Arc<AppState>, factor: f64) {
-        if let Some(browser) = self.live_browser(state, block_id) {
-            if let Some(frame) = browser.main_frame() {
-                let js = format!("document.documentElement.style.zoom = '{factor}';");
-                let code = CefString::from(js.as_str());
-                let url = CefString::from("");
-                frame.execute_java_script(Some(&code), Some(&url), 0);
-            }
-        }
-    }
-
-    /// Re-apply this pane's stored zoom (if it's ever been changed from the
-    /// 1.0 default) after a navigation — a fresh page load replaces the
-    /// previous page's DOM/inline-style state entirely, so the CSS `zoom`
-    /// injected before that navigation is gone along with it. Called from
-    /// `browser_pane::callbacks::on_load_end_browser_pane`, which already
-    /// fires after every navigation for other per-pane HWND bookkeeping.
-    /// A no-op for panes that were never zoomed (no map entry) — every
-    /// pane's first ever load doesn't pay for a wasted execute_java_script
-    /// call setting zoom to its already-default 1.0.
-    pub fn reapply_zoom(&self, block_id: &str, state: &Arc<AppState>) {
-        let factor = state.browser_pane_zoom.lock().get(block_id).copied();
-        if let Some(factor) = factor {
-            self.apply_zoom(block_id, state, factor);
-        }
-    }
-
-    /// Tell every live pane browser it has lost focus, at the Chromium level.
-    /// Panes in `Closing` are skipped — their HWND may be mid-destruction and
-    /// `set_focus(0)` against it can hit an invalid render widget.
-    pub fn defocus_all(&self, state: &Arc<AppState>) {
-        // Phase H.1.b + H.2.b — read live labels via reducer-aware helper,
-        // then look up each browser via reducer-aware helper. Both with
-        // fallback + drift logging.
-        let labels = state.live_browser_pane_labels();
-        for label in &labels {
-            if let Some(browser) = state.get_browser(label) {
-                if let Some(host) = browser.host() {
-                    host.set_focus(0);
-                }
-            }
-        }
-    }
-
-    /// Apply a clip region to every live pane HWND that subtracts the given
-    /// overlay rects (in main-window client coordinates). The pane renders
-    /// normally outside the overlay region; inside it, the HWND is
-    /// transparent so the DOM overlay painted at the same screen position
-    /// shows through.
-    ///
-    /// This is the Win32 "airspace" workaround — native HWNDs always paint
-    /// above DOM regardless of CSS z-index, and `SetWindowRgn` is the one
-    /// mechanism that lets DOM bleed through a specific region of a child
-    /// HWND. Empty `overlay_rects` restores full pane visibility (same as
-    /// calling `clear_pane_overlay_clip`).
-    ///
-    /// No-op on non-Windows: other platforms don't use native child HWNDs
-    /// for panes, so there's no airspace to work around.
-    ///
-    /// `window_label` scopes the clip to panes whose top-level ancestor
-    /// matches the requesting window. Without it, a modal opened in
-    /// window B would clip panes in window A (see Codex P1 on PR #544).
-    /// Empty string matches today's legacy callers that don't know their
-    /// window label — falls through to the no-filter behaviour for
-    /// back-compat until every caller is updated.
-    #[cfg(target_os = "windows")]
-    pub fn set_pane_overlay_clip(
-        &self,
-        state: &Arc<AppState>,
-        window_label: &str,
-        overlay_rects: &[(i32, i32, i32, i32)],
-    ) {
-        use windows_sys::Win32::Foundation::{POINT, RECT};
-        use windows_sys::Win32::Graphics::Gdi::{
-            CombineRgn, CreateRectRgn, DeleteObject, InvalidateRect, MapWindowPoints, SetWindowRgn,
-            RGN_DIFF,
-        };
-        use windows_sys::Win32::UI::WindowsAndMessaging::{
-            GetAncestor, GetWindowRect, GA_ROOT,
-        };
-
-        // Resolve the requesting window's top-level HWND so we can filter
-        // panes by ownership. If the label is unknown we fall through with
-        // no filter — matches pre-scoping behaviour rather than silently
-        // doing nothing.
-        let requesting_top_level: *mut std::ffi::c_void = if window_label.is_empty() {
-            std::ptr::null_mut()
-        } else {
-            // Phase H.2.b — reducer-aware lookup with fallback.
-            match state.get_browser(window_label).and_then(|b| b.host()) {
-                Some(host) => {
-                    let h = host.window_handle();
-                    if h.0.is_null() {
-                        std::ptr::null_mut()
-                    } else {
-                        unsafe { GetAncestor(h.0 as _, GA_ROOT) as *mut std::ffi::c_void }
-                    }
-                }
-                None => std::ptr::null_mut(),
-            }
-        };
-
-        // Phase H.1.b + H.2.b — labels via reducer-aware helper; per-label
-        // browser lookup via reducer-aware helper. Drops the held-across-loop
-        // legacy lock; each iteration now snapshots independently.
-        let labels = state.live_browser_pane_labels();
-        for label in &labels {
-            let browser = match state.get_browser(label) {
-                Some(b) => b,
-                None => continue,
-            };
-            let host = match browser.host() {
-                Some(h) => h,
-                None => continue,
-            };
-            let hwnd_raw = host.window_handle();
-            if hwnd_raw.0.is_null() {
-                continue;
-            }
-            let pane_hwnd = hwnd_raw.0 as *mut std::ffi::c_void;
-
-            // Window-scope filter. Skip panes whose top-level HWND differs
-            // from the requesting window's. `null` requesting = legacy
-            // caller / no-op filter (applies to all panes).
-            if !requesting_top_level.is_null() {
-                let pane_top = unsafe { GetAncestor(pane_hwnd as _, GA_ROOT) as *mut std::ffi::c_void };
-                if pane_top != requesting_top_level {
-                    continue;
-                }
-            }
-
-            // App-owned wrapper HWND (browser_pane::wrapper — see its module
-            // doc and SPEC_BROWSER_PANE_WINDOWS_TEARDOWN_SPIKE_2026_07_03.md)
-            // sits directly between `pane_hwnd` and its host window and is
-            // kept congruent with it (same origin, same size — its own
-            // `WM_SIZE` handler resizes CEF's child to exactly fill it, see
-            // `wrapper.rs`). `SetWindowRgn` only changes hit-testing for the
-            // HWND it's called on: clipping `pane_hwnd` alone stops IT from
-            // painting/claiming the hole, but Win32 hit-testing for a point
-            // inside that hole then resolves to `pane_hwnd`'s own parent —
-            // the wrapper — which still claims its full, unclipped
-            // rectangle there. The wrapper's window class paints nothing of
-            // its own (`hbrBackground: null`) so the DOM underneath still
-            // shows through visually (matches "renders fine"), but its
-            // `WM_LBUTTONDOWN`/`WM_MOUSEMOVE` fall through to
-            // `DefWindowProcW` unhandled — the input is silently absorbed
-            // there instead of reaching the DOM (matches "hover/click
-            // no-op"). See docs/analysis/ANALYSIS_WINDOWS_PANE_OVERLAY_WRAPPER_HITTEST_GAP_2026_07_13.md.
-            // Every region applied to `pane_hwnd` below must therefore also
-            // be applied to the wrapper, in the same wrapper-local
-            // coordinates (congruent, so no separate coordinate math is
-            // needed). `None` only if the wrapper hasn't been created yet /
-            // already torn down (a real race, not the common case — every
-            // embedded pane creation on Windows creates one unconditionally,
-            // `browser_pane/creation.rs`); degrade gracefully by clipping
-            // just the pane in that case rather than failing the whole call.
-            let wrapper_hwnd = crate::browser_pane::wrapper::peek_wrapper_hwnd(label)
-                .map(|h| h as *mut std::ffi::c_void);
-
-            unsafe {
-                // Empty overlay list = restore full visibility (region=NULL).
-                // bRedraw=FALSE (0) — see #1097 fix #5. With bRedraw=TRUE
-                // SetWindowRgn does a SYNCHRONOUS redraw inside the call
-                // (it sends WM_WINDOWPOSCHANGING/CHANGED + a paint pass),
-                // which serializes against the IPC handler. Passing FALSE
-                // updates the region without painting; an explicit
-                // InvalidateRect schedules WM_PAINT for the NEXT message
-                // pump tick. For a menu hover that re-fires this code path
-                // several times per gesture, that's a material win.
-                if overlay_rects.is_empty() {
-                    // Skip if this pane is already at full visibility (same HWND,
-                    // already restored) — avoids a redundant SetWindowRgn +
-                    // repaint every call while any overlay is open elsewhere.
-                    //
-                    // The cache guard is held across check → apply → record so
-                    // the sequence is atomic. This handler runs INLINE on a
-                    // multi-threaded tokio worker (ipc.rs — no UI-thread marshal,
-                    // unlike the non-Windows path), so two concurrent clips for
-                    // the same pane could otherwise interleave their SetWindowRgn
-                    // vs cache.insert and leave the recorded sig disagreeing with
-                    // the live region — a later wrong-skip.
-                    let sig = clip_sig_whole(pane_hwnd as isize);
-                    let mut cache = state.pane_clip_cache.lock();
-                    if cache.get(label).copied() == Some(sig) {
-                        continue;
-                    }
-                    let applied = SetWindowRgn(pane_hwnd as _, std::ptr::null_mut(), 0) != 0;
-                    // The pane is being made WHOLE again — the previously
-                    // hidden sub-area needs to repaint its content. We
-                    // don't track the diff yet (Phase 2 follow-up), so
-                    // invalidate the entire pane; CEF will paint only
-                    // dirty regions on the next pump.
-                    InvalidateRect(pane_hwnd as _, std::ptr::null(), 0);
-                    // Restore the wrapper's region too — see the wrapper_hwnd
-                    // comment above. Not gated on `applied`/cached separately:
-                    // the wrapper and pane_hwnd always change in lockstep, so
-                    // the pane's own sig is the correct gate for both.
-                    if let Some(wrapper_hwnd) = wrapper_hwnd {
-                        SetWindowRgn(wrapper_hwnd as _, std::ptr::null_mut(), 0);
-                        InvalidateRect(wrapper_hwnd as _, std::ptr::null(), 0);
-                    }
-                    // Only record the sig if the region actually applied; on a
-                    // (rare) SetWindowRgn failure leave the entry unset so the
-                    // next call retries instead of skipping a clip that never
-                    // landed.
-                    if applied {
-                        cache.insert(label.clone(), sig);
-                    }
-                    continue;
-                }
-
-                // Resolve the pane's position in its parent (main window)
-                // client coords so we can translate overlay rects (which
-                // arrive in main-window client coords from the frontend)
-                // into pane-local coords for the region API.
-                //
-                // GA_ROOT, not single-hop GetParent: since
-                // SPEC_BROWSER_PANE_WINDOWS_TEARDOWN_SPIKE_2026_07_03.md,
-                // `pane_hwnd` (CEF's own HWND) is a WS_CHILD of our own
-                // wrapper HWND, not directly of main — GetParent would
-                // resolve to the wrapper (whose client coords are always
-                // (0,0)-origin at the pane's own size), not main, silently
-                // breaking this translation. GA_ROOT walks past any number
-                // of intermediate layers to the actual top-level window,
-                // which is what "main window client coords" means here —
-                // same pattern already used above for requesting_top_level.
-                let parent = GetAncestor(pane_hwnd as _, GA_ROOT);
-                if parent.is_null() {
-                    continue;
-                }
-                let mut pane_rect: RECT = std::mem::zeroed();
-                if GetWindowRect(pane_hwnd as _, &mut pane_rect) == 0 {
-                    continue;
-                }
-                // Convert pane_rect from screen coords to parent client
-                // coords by mapping its two corner points.
-                let pts_ptr = &mut pane_rect as *mut RECT as *mut POINT;
-                MapWindowPoints(std::ptr::null_mut(), parent, pts_ptr, 2);
-
-                let pane_w = pane_rect.right - pane_rect.left;
-                let pane_h = pane_rect.bottom - pane_rect.top;
-                if pane_w <= 0 || pane_h <= 0 {
-                    continue;
-                }
-
-                // Skip if the region we'd compute is identical to the one
-                // already applied to this HWND (same geometry + same overlays).
-                //
-                // As in the whole-visibility branch above, the cache guard is
-                // held across check → build → apply → record so a concurrent
-                // clip for the same pane (this handler runs inline on a
-                // multi-threaded tokio worker) can't interleave SetWindowRgn and
-                // cache.insert and desync the recorded sig from the live region.
-                let sig = clip_sig_clipped(
-                    pane_hwnd as isize,
-                    pane_rect.left,
-                    pane_rect.top,
-                    pane_w,
-                    pane_h,
-                    overlay_rects,
-                );
-                let mut cache = state.pane_clip_cache.lock();
-                if cache.get(label).copied() == Some(sig) {
-                    continue;
-                }
-
-                // Build a region in pane-local coords: start with full pane,
-                // subtract every overlay rect that intersects it. Factored
-                // into a closure because it must run TWICE — once per
-                // target HWND — since `SetWindowRgn` takes ownership of the
-                // region handle it's given on success, so the same GDI
-                // region object can't be handed to two different windows.
-                // The wrapper is congruent with `pane_hwnd` (same origin,
-                // same size — see the wrapper_hwnd comment above), so the
-                // exact same `pane_w`/`pane_h`/`pane_rect` inputs apply to
-                // both without any extra coordinate math.
-                let build_region = || -> *mut std::ffi::c_void {
-                    let region = CreateRectRgn(0, 0, pane_w, pane_h);
-                    if region.is_null() {
-                        return region;
-                    }
-                    for (ox, oy, ow, oh) in overlay_rects {
-                        // Translate overlay rect (window client coords) →
-                        // pane-local coords by subtracting pane's window pos.
-                        let left = ox - pane_rect.left;
-                        let top = oy - pane_rect.top;
-                        let right = left + ow;
-                        let bottom = top + oh;
-                        // Skip if no intersection with the pane's local bounds.
-                        if right <= 0 || bottom <= 0 || left >= pane_w || top >= pane_h {
-                            continue;
-                        }
-                        let overlay_rgn = CreateRectRgn(left, top, right, bottom);
-                        if !overlay_rgn.is_null() {
-                            CombineRgn(region, region, overlay_rgn, RGN_DIFF);
-                            DeleteObject(overlay_rgn as _);
-                        }
-                    }
-                    region
-                };
-
-                let region = build_region();
-                if region.is_null() {
-                    continue;
-                }
-                // SetWindowRgn takes ownership of the region handle on
-                // success; the system frees it when the window is destroyed
-                // or a new region is set. On FAILURE ownership stays with us,
-                // so we DeleteObject the region below to avoid a GDI leak.
-                // bRedraw=FALSE — async paint via InvalidateRect below.
-                // See the empty-overlay branch above for the rationale.
-                let applied = SetWindowRgn(pane_hwnd as _, region as _, 0) != 0;
-                // The clip may have GROWN (more area hidden) or SHRUNK
-                // (less area hidden); we don't track the diff yet, so
-                // invalidate the whole pane. CEF dirty-region painting
-                // keeps the actual GPU work proportional to the change.
-                InvalidateRect(pane_hwnd as _, std::ptr::null(), 0);
-                // Same clip on the wrapper — see the wrapper_hwnd comment
-                // above for why this is required, not optional, for
-                // hover/click to reach the DOM inside the punched hole. A
-                // second, independently-created region handle (never the
-                // same handle passed to pane_hwnd above — see the ownership
-                // note). Best-effort: a wrapper-side failure here doesn't
-                // block the pane's own clip from being recorded, matching
-                // the "degrade gracefully" stance for a missing wrapper_hwnd
-                // above.
-                if let Some(wrapper_hwnd) = wrapper_hwnd {
-                    let wrapper_region = build_region();
-                    if !wrapper_region.is_null() {
-                        if SetWindowRgn(wrapper_hwnd as _, wrapper_region as _, 0) == 0 {
-                            DeleteObject(wrapper_region as _);
-                        }
-                        InvalidateRect(wrapper_hwnd as _, std::ptr::null(), 0);
-                    }
-                }
-                // Only record the sig when the region actually applied (matches
-                // the whole-visibility branch); on failure free the orphaned
-                // region and leave the entry unset so the next call retries.
-                if applied {
-                    cache.insert(label.clone(), sig);
-                } else {
-                    DeleteObject(region as _);
-                }
-            }
-        }
-        // Prune cache entries for panes no longer live (closed/destroyed),
-        // bounding the map. Labels are never reused for a different pane, so a
-        // dropped entry only means the next clip for a (re)created pane applies
-        // fresh — which is correct.
-        {
-            let mut cache = state.pane_clip_cache.lock();
-            cache.retain(|k, _| labels.iter().any(|l| l == k));
-        }
-        tracing::info!(
-            pane_count = labels.len(),
-            overlay_count = overlay_rects.len(),
-            "[pane-airspace] applied overlay clip to pane HWNDs",
-        );
-    }
-    /// Linux/macOS — equivalent of the Windows SetWindowRgn airspace
-    /// workaround, but built on Views instead of HWND clip regions.
-    ///
-    /// `add_overlay_view` puts the pane on a higher z-layer than the host UI
-    /// BrowserView, so any host-side modal/dropdown/contextmenu that overlaps
-    /// a pane rect renders UNDERNEATH the pane and becomes unclickable. We
-    /// can't punch a clip hole through an Aura View the way Win32 SetWindowRgn
-    /// does on an HWND. The pragmatic workaround: when ANY overlay rect
-    /// overlaps a pane's bounds, hide that pane (`set_visible(false)`); when
-    /// no overlay rect intersects, show it again. The DOM modal renders in
-    /// the host UI BrowserView underneath, becomes the topmost paint at that
-    /// rect, and the pane's content briefly disappears — same UX trade-off
-    /// the Windows path makes (Win32 punches a hole; we hide the whole pane).
-    /// (Codex P1 on PR #682.)
-    ///
-    /// Future improvement: only hide the overlapping fraction of the pane
-    /// (would need a per-overlay set_size + position trick or a custom Layout).
-    /// Hiding the whole pane is acceptable for now — the airspace problem only
-    /// arises when modals open over panes, which is a transient case.
-    ///
-    /// `_window_label` is currently ignored on this path because we only
-    /// support a single primary window for panes (sub-window panes are a
-    /// follow-up — see PR #682's "Risks / follow-ups" section).
-    /// (See doc comment for the cfg(target_os = "windows") variant.)
-    /// Linux/macOS body — marshalled to the CEF UI thread because
-    /// `OverlayController::set_visible` and `bounds()` are UI-thread-only.
-    /// IPC handler runs on tokio so we post a task and return immediately.
-    /// `window_label` filters which panes get visibility-managed: only those
-    /// attached to the requesting window are affected.
-    #[cfg(not(target_os = "windows"))]
-    pub fn set_pane_overlay_clip(
-        &self,
-        state: &Arc<AppState>,
-        window_label: &str,
-        overlay_rects: &[(i32, i32, i32, i32)],
-    ) {
-        // Publish to AppState so resize_browser_pane_view can consult the
-        // same authoritative rect list when computing pane visibility on
-        // its own code path. Without this, a positive-dimension resize
-        // (e.g. user drags a splitter while a DOM modal is open) would
-        // call set_visible(1) and re-expose the pane on top of the modal.
-        // See state::pane_overlay_rects doc comment.
-        state
-            .pane_overlay_rects
-            .lock()
-            .insert(window_label.to_string(), overlay_rects.to_vec());
-
-        let mut task = SetPaneOverlayClipViewsTask::new(
-            state.clone(),
-            window_label.to_string(),
-            overlay_rects.to_vec(),
-        );
-        cef::post_task(cef::ThreadId::UI, Some(&mut task));
-    }
-
-    /// Give keyboard focus to the pane's child HWND so keystrokes reach the
-    /// embedded page. Called by the frontend's ViewModel.giveFocus() when the
-    /// pane becomes the active layout node — without this, focus falls back to
-    /// the main window's invisible "dummy-focus" input and keystrokes vanish.
-    ///
-    /// No-ops if the pane is `Closing`: a SetFocus against a HWND that CEF is
-    /// concurrently tearing down is the exact race documented in
-    /// `SPEC_BROWSER_PANE_LIFECYCLE.md` §5 race #2.
-    pub fn focus(&self, block_id: &str, state: &Arc<AppState>) {
-        if let Some(browser) = self.live_browser(state, block_id) {
-            if let Some(host) = browser.host() {
-                host.set_focus(1);
-                #[cfg(target_os = "windows")]
-                {
-                    let hwnd = host.window_handle();
-                    if !hwnd.0.is_null() {
-                        // Tell the subclass this focus request is intentional
-                        // (not Chromium's on-load focus steal) so it won't be
-                        // redirected back to the parent.
-                        crate::browser_pane::ALLOW_BROWSER_PANE_FOCUS_ONCE.store(
-                            true,
-                            std::sync::atomic::Ordering::Relaxed,
-                        );
-                        unsafe {
-                            windows_sys::Win32::UI::Input::KeyboardAndMouse::SetFocus(hwnd.0 as _);
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Pure clamp+step arithmetic for `BrowserPaneManager::step_zoom`, split out
-/// so it's testable without a real `AppState`/lock/browser — mirrors why
-/// `BrowserPaneCloseOps` exists as its own trait (see that trait's doc
-/// comment above): the CEF/HWND side can't be unit-tested meaningfully, but
-/// the logic feeding it can.
-fn next_zoom_factor(current: f64, delta: f64) -> f64 {
-    (current + delta).clamp(0.5, 2.0)
-}
-
-// `CreateBrowserPaneTask` moved to `crate::browser_pane::creation` in Phase 3.
-
-/// Two axis-aligned rects intersect iff neither is fully to one side of the
-/// other. Coordinates: (x, y, width, height). Used by the Linux/macOS
-/// pane-airspace logic to decide whether an overlay rect from the frontend
-/// covers any part of a pane's bounds.
-#[cfg(not(target_os = "windows"))]
-fn rects_intersect(a: (i32, i32, i32, i32), b: (i32, i32, i32, i32)) -> bool {
-    let (ax, ay, aw, ah) = a;
-    let (bx, by, bw, bh) = b;
-    let a_right = ax + aw;
-    let a_bottom = ay + ah;
-    let b_right = bx + bw;
-    let b_bottom = by + bh;
-    !(a_right <= bx || b_right <= ax || a_bottom <= by || b_bottom <= ay)
-}
-
-/// Compute whether a pane with the given bounds should be visible, given
-/// the pane's parent window. Both pane-airspace (`SetPaneOverlayClipViewsTask`)
-/// and per-pane resize (`resize_browser_pane_view`) call this to converge on
-/// the same answer — without it, the two paths fight each other (Codex
-/// review on PR #881 caught the dragging-splitter-while-modal-open case
-/// where a positive resize re-exposed a pane that airspace had hidden).
-///
-/// A pane is visible iff BOTH conditions hold:
-/// - Its rect has non-zero width and height (frontend places it in a
-///   `display:none` placeholder when the tab is inactive → reports 0×0).
-/// - It does not intersect any registered overlay-clip rect for its window
-///   (e.g. a hamburger menu, tooltip, modal popover).
-// Linux-only since the macOS hole-punch mask (ui_tasks/pane_hole_mask.rs):
-// macOS paths use rect-only visibility and mask overlays instead of hiding.
-#[cfg(all(not(target_os = "windows"), not(target_os = "macos")))]
-pub fn compute_pane_visible(
-    state: &Arc<AppState>,
-    window_label: &str,
-    pane_rect: (i32, i32, i32, i32),
-) -> bool {
-    let (_, _, w, h) = pane_rect;
-    if w <= 0 || h <= 0 {
-        return false;
-    }
-    let rects = state.pane_overlay_rects.lock();
-    let overlays = match rects.get(window_label) {
-        Some(v) => v.clone(),
-        None => return true,
-    };
-    drop(rects);
-    !overlays.iter().any(|or| rects_intersect(*or, pane_rect))
-}
-
-// ── UI-thread marshalling tasks ─────────────────────────────────────────────
-//
-// `View::set_bounds` and `Window::remove_child_view` must run on the CEF UI
-// thread. Both `BrowserPaneManager::resize` and `::close` are called from IPC
-// handler tasks on tokio threads, so we wrap the UI-thread bodies in
-// `wrap_task!` structs and post them via `post_task(ThreadId::UI, ...)` —
-// same pattern as `ui_tasks::CloseWindowTask` / `MaximizeWindowTask` / etc.
-//
-// The same constraint applies to the Windows path's Win32 calls:
-// `DestroyWindow` only works from the thread that created the window (the CEF
-// UI thread, via CreateBrowserPaneTask) — see `DestroyPaneWrapperTask` below
-// and retro-browser-pane-renderer-leak-2026-07-07 for the leak that calling
-// it from the IPC thread silently caused.
-
-/// Windows pane-wrapper destroy, marshalled to the CEF UI thread (the
-/// wrapper's owning thread — Win32 DestroyWindow is owner-thread-only).
-/// Posted by `AppStateCloseOps::destroy_hwnd`; the actual
-/// hide → reparent-to-top-level → DestroyWindow sequence lives in
-/// `browser_pane::wrapper::destroy_wrapper_hwnd`.
-#[cfg(target_os = "windows")]
-wrap_task! {
-    pub struct DestroyPaneWrapperTask {
-        hwnd: isize,
-    }
-
-    impl Task {
-        fn execute(&self) {
-            crate::browser_pane::wrapper::destroy_wrapper_hwnd(
-                self.hwnd as *mut std::ffi::c_void,
-            );
-        }
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-wrap_task! {
-    pub struct ResizeBrowserPaneViewTask {
-        state: Arc<AppState>,
-        label: String,
-        rect: Rect,
-    }
-
-    impl Task {
-        fn execute(&self) {
-            crate::browser_pane::creation_views::resize_browser_pane_view(
-                &self.state, &self.label, self.rect.clone(),
-            );
-        }
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-wrap_task! {
-    pub struct DetachBrowserPaneViewTask {
-        state: Arc<AppState>,
-        label: String,
-    }
-
-    impl Task {
-        fn execute(&self) {
-            crate::browser_pane::creation_views::detach_browser_pane_view(
-                &self.state, &self.label,
-            );
-        }
-    }
-}
-
-/// Linux/macOS pane-airspace task — fired by `set_pane_overlay_clip` for the
-/// non-Windows code path. For each live OverlayController, hide it when any
-/// overlay rect intersects its current bounds; show it otherwise. See the
-/// doc comment on `set_pane_overlay_clip` (non-Windows variant) for why this
-/// is the equivalent of the Windows SetWindowRgn airspace dance.
-#[cfg(not(target_os = "windows"))]
-wrap_task! {
-    pub struct SetPaneOverlayClipViewsTask {
-        state: Arc<AppState>,
-        // Only panes attached to this window get visibility-managed; panes
-        // in other windows are unaffected by overlay rects from this window.
-        // Mirrors the window_label filtering in the Windows path.
-        window_label: String,
-        overlay_rects: Vec<(i32, i32, i32, i32)>,
-    }
-
-    impl Task {
-        fn execute(&self) {
-            // Snapshot the (pane label, parent window label, controller) tuples,
-            // filter by parent-window-label matching the requesting window,
-            // drop the mutex before any FFI call (snapshot-and-drop discipline
-            // per docs/specs/SPEC_PHASE_F_HOST_REDUCER §6).
-            let live: Vec<(String, cef::OverlayController)> = self
-                .state
-                .browser_pane_overlays
-                .lock()
-                .iter()
-                .filter(|(_, (win_label, _))| win_label == &self.window_label)
-                .map(|(k, (_, c))| (k.clone(), c.clone()))
-                .collect();
-            if live.is_empty() {
-                return;
-            }
-            for (label, controller) in live {
-                use cef::ImplOverlayController;
-                // On macOS, `controller.bounds()` cannot be trusted here — CEF
-                // Views' own set_size/set_position are permanent no-ops on
-                // NativeWidgetMacNSWindow, so bounds() reflects a stale,
-                // wrong-scale (DIP, not physical-px) rect rather than the real
-                // on-screen frame we committed via raw ObjC setFrame:. Using it
-                // made the intersection test silently miss real overlaps —
-                // visible stayed `true` while a DOM menu was drawn on top, so
-                // the pane kept intercepting clicks meant for the DOM even
-                // though it painted correctly underneath. Use the maintained
-                // physical-rect cache instead; see
-                // `AppState::browser_pane_physical_rects`. Fall back to
-                // bounds() if the cache somehow has no entry yet (shouldn't
-                // happen — creation seeds it before any overlay-clip can run).
-                #[cfg(target_os = "macos")]
-                {
-                    let pane_rect = self
-                        .state
-                        .browser_pane_physical_rects
-                        .lock()
-                        .get(&label)
-                        .copied()
-                        .unwrap_or_else(|| {
-                            let pb = controller.bounds();
-                            (pb.x, pb.y, pb.width, pb.height)
-                        });
-                    let (px, py, pw, ph) = pane_rect;
-                    if pw <= 0 || ph <= 0 {
-                        // Tab inactive (placeholder collapsed) — keep hidden.
-                        controller.set_visible(0);
-                        continue;
-                    }
-                    // Hole punch (Windows SetWindowRgn parity): mask out the
-                    // overlay∩pane rects instead of hiding the whole pane. The
-                    // pane stays LIVE and visible around the holes; the DOM
-                    // overlay shows through them, and transparent pixels
-                    // click-through to the main window. Whole-pane hide (and
-                    // the frontend freeze-frame that compensated for it) only
-                    // remains as the fallback when the overlay NSWindow can't
-                    // be found. See ui_tasks/pane_hole_mask.rs.
-                    let holes: Vec<(i32, i32, i32, i32)> = self
-                        .overlay_rects
-                        .iter()
-                        .filter(|or| rects_intersect(**or, pane_rect))
-                        .map(|&(ox, oy, ow, oh)| {
-                            let hx = ox.max(px);
-                            let hy = oy.max(py);
-                            let hr = (ox + ow).min(px + pw);
-                            let hb = (oy + oh).min(py + ph);
-                            (hx, hy, hr - hx, hb - hy)
-                        })
-                        .collect();
-                    let wnum = self
-                        .state
-                        .browser_pane_overlay_wnums
-                        .lock()
-                        .get(&label)
-                        .copied()
-                        .unwrap_or(0);
-                    let masked = crate::ui_tasks::pane_hole_mask::apply_pane_overlay_hole_mask(
-                        wnum, pane_rect, &holes,
-                    );
-                    if masked {
-                        controller.set_visible(1);
-                    } else {
-                        // Fallback: window lookup failed — old whole-pane hide.
-                        let visible = holes.is_empty();
-                        controller.set_visible(if visible { 1 } else { 0 });
-                    }
-                    tracing::debug!(
-                        label = %label,
-                        window_label = %self.window_label,
-                        masked, hole_count = holes.len(),
-                        pane_x = px, pane_y = py, pane_w = pw, pane_h = ph,
-                        overlay_count = self.overlay_rects.len(),
-                        "[pane-airspace] views: applied hole mask"
-                    );
-                }
-                #[cfg(not(target_os = "macos"))]
-                {
-                    let pane_rect = {
-                        let pb = controller.bounds();
-                        (pb.x, pb.y, pb.width, pb.height)
-                    };
-                    // Shared visibility helper consults BOTH the pane's own rect
-                    // (zero → hidden because tab inactive) and the latest
-                    // overlay-clip rects published in AppState. Resize path uses
-                    // the same helper so both decisions converge.
-                    let visible = compute_pane_visible(&self.state, &self.window_label, pane_rect);
-                    controller.set_visible(if visible { 1 } else { 0 });
-                    tracing::debug!(
-                        label = %label,
-                        window_label = %self.window_label,
-                        visible,
-                        pane_x = pane_rect.0, pane_y = pane_rect.1, pane_w = pane_rect.2, pane_h = pane_rect.3,
-                        overlay_count = self.overlay_rects.len(),
-                        "[pane-airspace] views: applied visibility"
-                    );
-                }
-            }
-        }
-    }
-}
-
-// ── Tests ───────────────────────────────────────────────────────────────────
-//
-// Phase H.1.d/e (PR #5): The pane state machine lives in the host reducer
-// (`HostState.browser_panes`). Lifecycle transition tests — Live→Closing, idempotent
-// no-ops for missing or already-Closing entries, label sequence monotonicity,
-// drain-by-label — are now in `crate::reducer::tests`.
-//
-// What remains here: the FFI seam. `close_with` only takes a label and
-// drives `BrowserPaneCloseOps`; tests verify it forwards label → take → destroy
-// in order, with a None-returning `take` short-circuiting the destroy.
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-
-    /// Recording mock for `BrowserPaneCloseOps`. Tests inspect `taken` and
-    /// `destroyed` to assert what close_with did.
-    struct MockCloseOps {
-        registered: parking_lot::Mutex<HashMap<String, usize>>,
-        taken: parking_lot::Mutex<Vec<String>>,
-        destroyed: parking_lot::Mutex<Vec<usize>>,
-    }
-
-    impl MockCloseOps {
-        fn new() -> Self {
-            Self {
-                registered: parking_lot::Mutex::new(HashMap::new()),
-                taken: parking_lot::Mutex::new(Vec::new()),
-                destroyed: parking_lot::Mutex::new(Vec::new()),
-            }
-        }
-
-        fn register(&self, label: &str, hwnd: usize) {
-            self.registered.lock().insert(label.to_string(), hwnd);
-        }
-
-        fn taken_labels(&self) -> Vec<String> {
-            self.taken.lock().clone()
-        }
-
-        fn destroyed_hwnds(&self) -> Vec<usize> {
-            self.destroyed.lock().clone()
-        }
-    }
-
-    impl BrowserPaneCloseOps for MockCloseOps {
-        fn take_browser_hwnd(&self, label: &str) -> Option<usize> {
-            self.taken.lock().push(label.to_string());
-            self.registered.lock().remove(label)
-        }
-
-        fn destroy_hwnd(&self, hwnd: usize) {
-            self.destroyed.lock().push(hwnd);
-        }
-    }
-
-    #[test]
-    fn close_with_take_then_destroy_in_order() {
-        let ops = MockCloseOps::new();
-        ops.register("browser-pane-b1-1", 0xABCD);
-
-        BrowserPaneManager::close_with("browser-pane-b1-1", &ops);
-
-        assert_eq!(ops.taken_labels(), vec!["browser-pane-b1-1"]);
-        assert_eq!(ops.destroyed_hwnds(), vec![0xABCD]);
-    }
-
-    #[test]
-    fn close_with_no_hwnd_skips_destroy() {
-        // Browser was already gone (rare race — explicit close raced with
-        // an external close). take returns None; destroy must NOT be called.
-        let ops = MockCloseOps::new(); // no register() — lookup will miss
-
-        BrowserPaneManager::close_with("browser-pane-missing", &ops);
-
-        assert_eq!(ops.taken_labels(), vec!["browser-pane-missing"]);
-        assert!(ops.destroyed_hwnds().is_empty(),
-            "destroy_hwnd must not be called when take_browser_hwnd returns None");
-    }
-
-    #[test]
-    fn next_zoom_factor_steps_up_from_default() {
-        assert_eq!(super::next_zoom_factor(1.0, 0.05), 1.05);
-    }
-
-    #[test]
-    fn next_zoom_factor_steps_down_from_default() {
-        assert_eq!(super::next_zoom_factor(1.0, -0.05), 0.95);
-    }
-
-    #[test]
-    fn next_zoom_factor_clamps_at_max() {
-        assert_eq!(super::next_zoom_factor(2.0, 0.05), 2.0);
-        assert_eq!(super::next_zoom_factor(1.98, 0.05), 2.0);
-    }
-
-    #[test]
-    fn next_zoom_factor_clamps_at_min() {
-        assert_eq!(super::next_zoom_factor(0.5, -0.05), 0.5);
-        assert_eq!(super::next_zoom_factor(0.52, -0.05), 0.5);
-    }
-
-    #[test]
-    fn next_zoom_factor_many_steps_stay_in_bounds() {
-        let mut f = 1.0;
-        for _ in 0..200 {
-            f = super::next_zoom_factor(f, 0.05);
-        }
-        assert_eq!(f, 2.0);
-        for _ in 0..200 {
-            f = super::next_zoom_factor(f, -0.05);
-        }
-        assert_eq!(f, 0.5);
-    }
-}
